YP/requirements.txt
Django==5.1
aiohttp==3.10.5
APScheduler==3.10.4
django-apscheduler==0.6.2
pillow==10.4.0
requests==2.32.3
urllib3==2.2.2
celery==5.4.0
django-celery-beat==2.7.0
django-timezone-field==7.0
psycopg==3.2.3
loguru==0.7.2
transliterate==1.10.2
gunicorn==23.0.0
notifiers==1.3.3
django-tinymce==4.1.0
djangorestframework==3.16.0